+++ main.hpp
@@ -16,6 +16,8 @@ extern blit::Timer power_timer;
 extern uint8_t *level_tiles;
 extern uint8_t current_level;
 
+extern std::vector<int32_t> cycleTimes;
+
 extern Pacman *player;
 extern Blinky *blinky;
 extern Pinky *pinky;
@@ -27,6 +29,8 @@ extern uint32_t pills_eaten_this_life;
 
 extern bool operator==(blit::Point a, blit::Point b);
 extern bool operator!=(blit::Point a, blit::Point b);
+
+std::vector<int32_t> getCycleTimes();
 
 uint8_t level_get(blit::Point p);
 void level_set(blit::Point p, uint8_t e);